Day 2/Day 3/RAG/requirements.txt
langchain
langchain-community
langchain-google-genai
sentence-transformers
faiss-cpu
streamlit
pypdf
torch
python-dotenv
